+++ src/main/java/com/example/log4u/domain/supports/dto/SupportCreateRequestDto.java
@@ -1,5 +1,6 @@
 package com.example.log4u.domain.supports.dto;
 
+import com.example.log4u.domain.supports.supportType.SupportType;
 import jakarta.validation.constraints.NotBlank;
 import org.hibernate.validator.constraints.Length;
 
@@ -13,14 +14,4 @@ public record SupportCreateRequestDto(
         @NotBlank
         @Length(min = 2)
         String content
-) {
-    public enum SupportType {
-        TECHNICAL_ISSUE,
-        ACCOUNT_ISSUE,
-        PAYMENT_ISSUE,
-        FEATURE_REQUEST,
-        BILLING_ISSUE,
-        SECURITY_CONCERN,
-        ETC
-    }
-}
+) { }
