+++ src/main/java/com/example/log4u/domain/supports/supportType/SupportType.java
@@ -1,0 +1,11 @@
+package com.example.log4u.domain.supports.supportType;
+
+public enum SupportType {
+    TECHNICAL_ISSUE,
+    ACCOUNT_ISSUE,
+    PAYMENT_ISSUE,
+    FEATURE_REQUEST,
+    BILLING_ISSUE,
+    SECURITY_CONCERN,
+    ETC
+}
